requirements.txt
google-generativeai==0.7.2
anthropic>=0.39.0
boto3==1.34.144
pydantic==2.7.4
python-dotenv==1.0.1
requests==2.32.3
pypdf2==3.0.1
python-docx>=1.1.2
docx2pdf>=0.1.8 ; sys_platform == "win32"
python-jobspy>=1.1.79
pandas>=2.0.0